Move changes to `pnpm-runner` and rename it to `ci-runner`
