+++ eng/tools/ci-runner/test/actions.spec.js
@@ -5,7 +5,7 @@
 
 import { afterEach, assert, describe, it, vi } from "vitest";
 import { executeActions } from "../src/actions.js";
-import { spawnNode } from "../src/spawn.js";
+import { spawnNpx, spawnNpmRun } from "../src/spawn.js";
 import { getBaseDir } from "../src/env.js";
 import { join as pathJoin } from "node:path";
 
@@ -13,7 +13,8 @@ const baseDir = getBaseDir();
 
 vi.mock("../src/spawn.js", async () => {
   return {
-    spawnNode: vi.fn(),
+    spawnNpx: vi.fn(),
+    spawnNpmRun: vi.fn(),
   };
 });
 
@@ -22,63 +23,47 @@ describe("executeActions", () => {
     vi.clearAllMocks();
   });
 
-  it("should pass global actions directly to rush", () => {
+  it("should pass global actions directly to", () => {
     executeActions("unlink", [], [], "");
-    assert.deepEqual(vi.mocked(spawnNode).mock.calls, [
-      [baseDir, "common/scripts/install-run-rush.js", "unlink"],
-    ]);
+    assert.deepEqual(vi.mocked(spawnNpx).mock.calls, [[baseDir, "turbo", "run", "unlink"]]);
   });
 
   it("should run build commands for affected packages", () => {
     executeActions("build", ["appconfiguration"], [], "azure-app-configuration");
-    assert.deepEqual(vi.mocked(spawnNode).mock.calls, [
-      [
-        baseDir,
-        "common/scripts/install-run-rush.js",
-        "build",
-        "--from",
-        "@azure/app-configuration",
-      ],
+    assert.deepEqual(vi.mocked(spawnNpx).mock.calls, [
+      [baseDir, "turbo", "run", "build", "--filter=...@azure/app-configuration"],
     ]);
   });
 
   it("should run lint in appropriate folders", () => {
     executeActions("lint", ["appconfiguration"], [], "azure-app-configuration");
     const packageDir = pathJoin(baseDir, "sdk/appconfiguration/app-configuration");
-    const executeScript = pathJoin(baseDir, "common/scripts/install-run-rushx.js");
-    assert.deepEqual(vi.mocked(spawnNode).mock.calls, [[packageDir, executeScript, "lint"]]);
+    assert.deepEqual(vi.mocked(spawnNpmRun).mock.calls, [[packageDir, "lint"]]);
   });
 
-  it("should handle arbitrary rush commands", () => {
+  it("should handle arbitrary run commands", () => {
     executeActions("foo", ["appconfiguration"], [], "azure-app-configuration");
-    assert.deepEqual(vi.mocked(spawnNode).mock.calls, [
-      [baseDir, "common/scripts/install-run-rush.js", "foo", "--to", "@azure/app-configuration"],
+    assert.deepEqual(vi.mocked(spawnNpx).mock.calls, [
+      [baseDir, "turbo", "run", "foo", "--filter=@azure/app-configuration..."],
     ]);
   });
 
   it("should pass through arguments for rush", () => {
     const rushArgs = ["--verbose", "-p max"];
     executeActions("build", ["appconfiguration"], rushArgs, "azure-app-configuration");
-    assert.deepEqual(vi.mocked(spawnNode).mock.calls, [
-      [
-        baseDir,
-        "common/scripts/install-run-rush.js",
-        "build",
-        "--from",
-        "@azure/app-configuration",
-        ...rushArgs,
-      ],
+    assert.deepEqual(vi.mocked(spawnNpx).mock.calls, [
+      [baseDir, "turbo", "run", "build", "--filter=...@azure/app-configuration", ...rushArgs],
     ]);
   });
 
   it("should return a non-zero return when a command fails", () => {
-    vi.mocked(spawnNode).mockReturnValueOnce(1);
+    vi.mocked(spawnNpx).mockReturnValueOnce(1);
     const resultCode = executeActions("build", ["appconfiguration"], [], "azure-app-configuration");
     assert.strictEqual(resultCode, 1);
   });
 
   it("should return a non-zero return when one of many commands fails", () => {
-    vi.mocked(spawnNode).mockReturnValueOnce(0).mockReturnValueOnce(1).mockReturnValueOnce(0);
+    vi.mocked(spawnNpmRun).mockReturnValueOnce(0).mockReturnValueOnce(1).mockReturnValueOnce(0);
     const resultCode = executeActions(
       "lint",
       ["appconfiguration", "storage", "keyvault"],
