+++ eng/tools/ci-runner/test/helpers.spec.js
@@ -1,0 +1,82 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+// @ts-check
+
+import { assert, describe, it } from "vitest";
+import {
+  getFilteredPackages,
+  reducedDependencyTestMatrix,
+  restrictedToPackages,
+} from "../src/helpers.js";
+
+describe("getDirectionMappedPackages", () => {
+  describe("build", () => {
+    it("should use reduced core scope for changed core package", () => {
+      const changed = ["@azure/core-client"];
+      const mapped = getFilteredPackages(changed, "build", ["core"]);
+      const expected = reducedDependencyTestMatrix["core"].map((p) =>
+        restrictedToPackages.includes(p) ? `--filter=${p}...` : `--filter=...${p}`,
+      );
+      expected.unshift("--filter=@azure/core-client...");
+
+      assert.deepStrictEqual(mapped, expected);
+    });
+
+    it("it uses --from when building a normal package", () => {
+      const changed = ["@azure/app-configuration"];
+      const mapped = getFilteredPackages(changed, "build", ["appconfiguration"]);
+
+      assert.deepStrictEqual(mapped, ["--filter=...@azure/app-configuration"]);
+    });
+
+    it("it uses --impacted-by when doing subsequent build tasks like build:test", () => {
+      const changed = ["@azure/storage-blob"];
+      const mapped = getFilteredPackages(changed, "build:test", ["storage"]);
+
+      assert.deepStrictEqual(mapped, ["--filter=...@azure/storage-blob"]);
+    });
+
+    it("should use --to and --from for mixed packages", () => {
+      const changed = ["@azure/core-rest-pipeline", "@azure/app-configuration"];
+      const mapped = getFilteredPackages(changed, "build", ["core", "appconfiguration"]);
+      const expected = reducedDependencyTestMatrix["core"].map((p) =>
+        restrictedToPackages.includes(p) ? `--filter=${p}...` : `--filter=...${p}`,
+      );
+
+      expected.unshift("--filter=...@azure/app-configuration");
+      expected.unshift("--filter=@azure/core-rest-pipeline...");
+
+      assert.deepStrictEqual(mapped, expected);
+    });
+  });
+
+  describe("test commands", () => {
+    it("should use the reduced dependency matrix plus --only when testing a core package", () => {
+      const changed = ["@azure/core-client"];
+      const mapped = getFilteredPackages(changed, "unit-test", ["core"]);
+      const expected = reducedDependencyTestMatrix["core"].map((p) => `--filter=${p}`);
+
+      expected.unshift("--filter=@azure/core-client");
+
+      assert.deepStrictEqual(mapped, expected);
+    });
+
+    it("it uses --impacted-by when testing a normal package", () => {
+      const changed = ["@azure/app-configuration"];
+      const mapped = getFilteredPackages(changed, "unit-test", ["appconfiguration"]);
+
+      assert.deepStrictEqual(mapped, ["--filter=...@azure/app-configuration"]);
+    });
+
+    it("should use --only when testing two or more service dirs", () => {
+      const changed = ["@azure/app-configuration", "@azure/storage-blob"];
+      const mapped = getFilteredPackages(changed, "unit-test", ["appconfiguration", "storage"]);
+
+      assert.deepStrictEqual(mapped, [
+        "--filter=@azure/app-configuration",
+        "--filter=@azure/storage-blob",
+      ]);
+    });
+  });
+});
