+++ eng/tools/ci-runner/vitest.config.mjs
@@ -2,11 +2,12 @@
 // Licensed under the MIT License.
 
 import { defineConfig } from "vitest/config";
+import { AzureSDKReporter } from "../../../vitest.shared.config.js";
 
 export default defineConfig({
   test: {
     testTimeout: 18000,
-    reporters: ["verbose"],
+    reporters: [new AzureSDKReporter()],
     watch: false,
     include: ["test/**/*.spec.js"],
     coverage: {
