+++ eng/tools/rush-runner/src/actions.js
@@ -3,59 +3,63 @@
 
 // @ts-check
 
-import { getFilteredPackages, getServicePackages, tryGetPkgRelativePath } from "./helpers.js";
-import { runAllWithDirection, runInPackageDirs, runGlobalAction } from "./rush.js";
+import {
+  getDirectionMappedPackages,
+  getServicePackages,
+  tryGetPkgRelativePath,
+} from "./helpers.js";
+import {
+  rushRunAll,
+  rushRunAllWithDirection,
+  runRushInPackageDirs,
+  rushGlobalAction,
+} from "./rush.js";
 
 /**
  *
  * @param {string} action - the command being performed
  * @param {string[]} serviceDirs - list of service directories impacted
- * @param {string[]} extraParams - commandline flags to pass
+ * @param {string[]} rushParams - commandline flags to pass directly to rush
  * @param {string} artifactNames - package names to filter to
  * @returns
  */
-export function executeActions(action, serviceDirs, extraParams, artifactNames, ciFlag) {
+export function executeActions(action, serviceDirs, rushParams, artifactNames, ciFlag) {
   const actionComponents = action.toLowerCase().split(":");
 
   console.log(`Packages to build: ${artifactNames}`);
   const { packageNames, packageDirs } = getServicePackages(serviceDirs, artifactNames);
-  console.log(`Packages eligible to run task: ${packageNames}`);
+  console.log(`Packages eligible to run rush task: ${packageNames}`);
 
   let exitCode = 0;
   if (serviceDirs.length === 0) {
-    exitCode = runGlobalAction(action, extraParams);
+    exitCode = rushGlobalAction(action, rushParams);
   } else {
     switch (actionComponents[0]) {
       case "build":
       case "test":
       case "unit-test":
       case "integration-test":
-        exitCode = runAllWithDirection(
+        exitCode = rushRunAllWithDirection(
           action,
-          getFilteredPackages(packageNames, action, serviceDirs),
-          extraParams,
-          ciFlag,
+          getDirectionMappedPackages(packageNames, action, serviceDirs),
+          rushParams,
+          ciFlag
         );
         break;
 
       case "lint":
-        exitCode = runInPackageDirs(action, packageDirs);
+        exitCode = runRushInPackageDirs(action, packageDirs);
         break;
       case "check-format":
-        exitCode = runInPackageDirs(action, packageDirs, (packageDir) => {
+        exitCode = runRushInPackageDirs(action, packageDirs, (packageDir) => {
           console.log(
-            `\nInvoke "npm run format" inside ${tryGetPkgRelativePath(packageDir)} to fix formatting\n`,
+            `\nInvoke "rushx format" inside ${tryGetPkgRelativePath(packageDir)} to fix formatting\n`,
           );
         });
         break;
 
       default:
-        exitCode = runAllWithDirection(
-          action,
-          packageNames.map((p) => `--filter=${p}...`),
-          extraParams,
-          ciFlag,
-        );
+        exitCode = rushRunAll(action, "--to", packageNames, rushParams);
         break;
     }
   }
