+++ eng/tools/rush-runner/src/helpers.js
@@ -59,17 +59,15 @@ export const restrictedToPackages = [
 ];
 
 /**
- * Helper function that determines the filter to use based on each individual package name
+ * Helper function that determines the rush command flag to use based on each individual package name
  *
  * If the targeted package is one of the restricted packages with a ton of dependents, we only want to run that package
  * and not all of its dependents.
  * @param {string[]} packageNames - An array of strings containing the packages names to run the action on.
  * @param {string} action - The action being performed ("build", "build:test", "build:samples", "unit-test:node", "unit-test:browser"
  * @param {string[]} serviceDirs - An array of strings containing the serviceDirs affected
- * @returns {string[]} - An array of workspace filters
  */
-export const getFilteredPackages = (packageNames, action, serviceDirs) => {
-  /** @type {string[]} */
+export const getDirectionMappedPackages = (packageNames, action, serviceDirs) => {
   const mappedPackages = [];
 
   let fullPackageNames = packageNames.slice();
@@ -92,26 +90,27 @@ export const getFilteredPackages = (packageNames, action, serviceDirs) => {
   if (action.startsWith("build")) {
     for (const packageName of fullPackageNames) {
       /**  @type {string} */
-      let filter;
+      let rushCommandFlag;
 
       if (restrictedToPackages.includes(packageName)) {
         // if this is one of our restricted packages with a ton of deps, make it targeted
         // as including all dependents will be too much
-        filter = `--filter=${packageName}...`; // package and its dependencies
+        rushCommandFlag = "--to";
       } else if (action === "build") {
-        filter = `--filter=...${packageName}`; // package and its dependents
+        rushCommandFlag = "--from";
       } else {
-        filter = `--filter=...${packageName}`;
+        // --impacted-by is only safe if the packages have already been built, since it won't build
+        // unrelated dependencies
+        rushCommandFlag = "--impacted-by";
       }
 
-      mappedPackages.push(filter);
+      mappedPackages.push([rushCommandFlag, packageName]);
     }
   } else {
-    mappedPackages.push(
-      ...fullPackageNames.map((p) =>
-        isReducedTestScopeEnabled ? `--filter=${p}` : `--filter=...${p}`,
-      ),
-    );
+    // we are in a test task of some kind
+    const rushCommandFlag = isReducedTestScopeEnabled ? "--only" : "--impacted-by";
+
+    mappedPackages.push(...fullPackageNames.map((p) => [rushCommandFlag, p]));
   }
 
   return mappedPackages;
