+++ eng/tools/rush-runner/src/spawn.js
@@ -5,64 +5,37 @@
 
 import { spawnSync } from "node:child_process";
 
-function isWindows() {
-  return process.platform === "win32";
-}
-
 /**
- * Helper function to spawn a command
+ * Helper function to spawn NodeJS programs
  *
- * @param {string} cmd - current working directory
  * @param {string} cwd - current working directory
  * @param {string[]} args - rest of arguments
  */
-function spawnWithLog(cmd, cwd, ...args) {
-  console.log(`Executing: "${cmd} ${args.join(" ")}" in ${cwd}\n\n`);
-  const proc = spawnSync(cmd, args, { cwd, stdio: "inherit", shell: isWindows() });
-  console.log(`\n\n${cmd} exited with code ${proc.status} `);
+export function spawnNode(cwd, ...args) {
+  console.log(`Executing: "node ${args.join(" ")}" in ${cwd}\n\n`);
+  const proc = spawnSync("node", args, { cwd, stdio: "inherit" });
+  console.log(`\n\nNode process exited with code ${proc.status} `);
 
   return proc.status ?? 1;
 }
 
 /**
- * Helper function to spawn npx to run programs
- *
- * @param {string} cwd - current working directory
- * @param {string[]} args - rest of arguments
- */
-export function spawnNpx(cwd, ...args) {
-  const command = isWindows() ? "npx.CMD" : "npx";
-  return spawnWithLog(command, cwd, ...args);
-}
-
-/**
- * Helper function to spawn `npm run` commands
- *
- * @param {string} cwd - current working directory
- * @param {string[]} args - rest of arguments
- */
-export function spawnNpmRun(cwd, ...args) {
-  const command = isWindows() ? "npm.CMD" : "npm";
-  return spawnWithLog(command, cwd, "run", ...args);
-}
-
-/**
- * Helper function to spawn npx to run programs and return the output
+ * Helper function to spawn NodeJS programs and return the output
  *
  * @param {string} cwd - current working directory
  * @param {string[]} args - rest of arguments
  * @returns {string} - output of the command
  */
-export function spawnNpxWithOutput(cwd, ...args) {
-  console.log(`Executing: "npx ${args.join(" ")}" in ${cwd}\n\n`);
-  const proc = spawnSync("npx", args, { cwd, stdio: "pipe" });
+export function spawnNodeWithOutput(cwd, ...args) {
+  console.log(`Executing: "node ${args.join(" ")}" in ${cwd}\n\n`);
+  const proc = spawnSync("node", args, { cwd, stdio: "pipe" });
 
   if (proc.error) {
     throw new Error(`Error executing command: ${proc.error.message}`);
   }
 
   const output = proc.stdout.toString();
-  console.log(`\n\nNpx exited with code ${proc.status}`);
+  console.log(`\n\nNode process exited with code ${proc.status}`);
 
   return output;
 }
