+++ eng/tools/rush-runner/test/helpers.spec.js
@@ -5,48 +5,48 @@
 
 import { assert, describe, it } from "vitest";
 import {
-  getFilteredPackages,
+  getDirectionMappedPackages,
   reducedDependencyTestMatrix,
   restrictedToPackages,
 } from "../src/helpers.js";
 
 describe("getDirectionMappedPackages", () => {
   describe("build", () => {
-    it("should use reduced core scope for changed core package", () => {
+    it("should use --to reduced core scope for changed core package", () => {
       const changed = ["@azure/core-client"];
-      const mapped = getFilteredPackages(changed, "build", ["core"]);
-      const expected = reducedDependencyTestMatrix["core"].map((p) =>
-        restrictedToPackages.includes(p) ? `--filter=${p}...` : `--filter=...${p}`,
-      );
-      expected.unshift("--filter=@azure/core-client...");
+      const mapped = getDirectionMappedPackages(changed, "build", ["core"]);
+      const expected = reducedDependencyTestMatrix["core"].map((p) => [
+        restrictedToPackages.includes(p) ? "--to" : "--from",
+        p,
+      ]);
+      expected.unshift(["--to", "@azure/core-client"]);
 
       assert.deepStrictEqual(mapped, expected);
     });
 
     it("it uses --from when building a normal package", () => {
       const changed = ["@azure/app-configuration"];
-      const mapped = getFilteredPackages(changed, "build", ["appconfiguration"]);
+      const mapped = getDirectionMappedPackages(changed, "build", ["appconfiguration"]);
 
-      assert.deepStrictEqual(mapped, ["--filter=...@azure/app-configuration"]);
+      assert.deepStrictEqual(mapped, [["--from", "@azure/app-configuration"]]);
     });
 
     it("it uses --impacted-by when doing subsequent build tasks like build:test", () => {
       const changed = ["@azure/storage-blob"];
-      const mapped = getFilteredPackages(changed, "build:test", ["storage"]);
+      const mapped = getDirectionMappedPackages(changed, "build:test", ["storage"]);
 
-      assert.deepStrictEqual(mapped, ["--filter=...@azure/storage-blob"]);
+      assert.deepStrictEqual(mapped, [["--impacted-by", "@azure/storage-blob"]]);
     });
 
     it("should use --to and --from for mixed packages", () => {
       const changed = ["@azure/core-rest-pipeline", "@azure/app-configuration"];
-      const mapped = getFilteredPackages(changed, "build", ["core", "appconfiguration"]);
-      const expected = reducedDependencyTestMatrix["core"].map((p) =>
-        restrictedToPackages.includes(p) ? `--filter=${p}...` : `--filter=...${p}`,
-      );
-
-      expected.unshift("--filter=...@azure/app-configuration");
-      expected.unshift("--filter=@azure/core-rest-pipeline...");
-
+      const mapped = getDirectionMappedPackages(changed, "build", ["core", "appconfiguration"]);
+      const expected = reducedDependencyTestMatrix["core"].map((p) => [
+        restrictedToPackages.includes(p) ? "--to" : "--from",
+        p,
+      ]);
+      expected.unshift(["--from", "@azure/app-configuration"]);
+      expected.unshift(["--to", "@azure/core-rest-pipeline"]);
       assert.deepStrictEqual(mapped, expected);
     });
   });
@@ -54,28 +54,30 @@ describe("getDirectionMappedPackages", () => {
   describe("test commands", () => {
     it("should use the reduced dependency matrix plus --only when testing a core package", () => {
       const changed = ["@azure/core-client"];
-      const mapped = getFilteredPackages(changed, "unit-test", ["core"]);
-      const expected = reducedDependencyTestMatrix["core"].map((p) => `--filter=${p}`);
-
-      expected.unshift("--filter=@azure/core-client");
+      const mapped = getDirectionMappedPackages(changed, "unit-test", ["core"]);
+      const expected = reducedDependencyTestMatrix["core"].map((p) => ["--only", p]);
+      expected.unshift(["--only", "@azure/core-client"]);
 
       assert.deepStrictEqual(mapped, expected);
     });
 
     it("it uses --impacted-by when testing a normal package", () => {
       const changed = ["@azure/app-configuration"];
-      const mapped = getFilteredPackages(changed, "unit-test", ["appconfiguration"]);
+      const mapped = getDirectionMappedPackages(changed, "unit-test", ["appconfiguration"]);
 
-      assert.deepStrictEqual(mapped, ["--filter=...@azure/app-configuration"]);
+      assert.deepStrictEqual(mapped, [["--impacted-by", "@azure/app-configuration"]]);
     });
 
     it("should use --only when testing two or more service dirs", () => {
       const changed = ["@azure/app-configuration", "@azure/storage-blob"];
-      const mapped = getFilteredPackages(changed, "unit-test", ["appconfiguration", "storage"]);
+      const mapped = getDirectionMappedPackages(changed, "unit-test", [
+        "appconfiguration",
+        "storage",
+      ]);
 
       assert.deepStrictEqual(mapped, [
-        "--filter=@azure/app-configuration",
-        "--filter=@azure/storage-blob",
+        ["--only", "@azure/app-configuration"],
+        ["--only", "@azure/storage-blob"],
       ]);
     });
   });
